Now collecting data from reddit. Displaying as part of the list and getting more at the bottom of the list.
Need to sort clicking the item.
Have not added a loading view in case loading times are longer than expected.

Still need to update error observable to show a error message

Need to check the get more is being called at the correct time, as we haven't updated the list size

Need to save state etc. Although we may not have time for this